+++ database/migrations/2018_05_26_121313_create_tests_table.php
@@ -1,0 +1,23 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class CreateTestsTable extends Migration
+{
+    public function up()
+    {
+        Schema::create('tests', function (Blueprint $table) {
+            $table->increments('id');
+            $table->string('name');
+            $table->integer('version');
+            $table->timestamps();
+        });
+    }
+
+    public function down()
+    {
+        Schema::dropIfExists('tests');
+    }
+}