+++ resources/assets/js/localState.js
@@ -1,0 +1,4 @@
+export default (context, state) => {
+    // initalise the local state
+};
+